feat: Selenium test implemented
