+++ app/modules/auth/tests/test_selenium_ide/test_cargarListado.py
@@ -1,18 +1,11 @@
-# Generated by Selenium IDE
-import pytest
-import time
-import json
 from selenium import webdriver
 from selenium.webdriver.common.by import By
-from selenium.webdriver.common.action_chains import ActionChains
-from selenium.webdriver.support import expected_conditions
-from selenium.webdriver.support.wait import WebDriverWait
-from selenium.webdriver.common.keys import Keys
-from selenium.webdriver.common.desired_capabilities import DesiredCapabilities
 from selenium.webdriver.chrome.service import Service
 from selenium.webdriver.chrome.options import Options
+from selenium.webdriver.common.keys import Keys
 
-class TestCargardataset():
+
+class TestCargarlistado:
     def setup_method(self, method):
         # Set up options for Chromium
         chrome_options = Options()
@@ -24,11 +17,11 @@ class TestCargardataset():
         # Initialize the WebDriver with the Chromium options
         self.driver = webdriver.Chrome(service=service, options=chrome_options)
         self.vars = {}
-  
+
     def teardown_method(self, method):
         self.driver.quit()
-  
-    def test_cargardatasets(self):
+
+    def test_cargarlistado(self):
         self.driver.get("http://127.0.0.1:5000/")
         self.driver.set_window_size(1854, 1048)
         self.driver.find_element(By.CSS_SELECTOR, ".nav-link:nth-child(1)").click()
@@ -37,8 +30,6 @@ class TestCargardataset():
         self.driver.find_element(By.ID, "password").click()
         self.driver.find_element(By.ID, "password").send_keys("281102")
         self.driver.find_element(By.ID, "password").send_keys(Keys.ENTER)
-        self.driver.find_element(By.LINK_TEXT, "List of users").click()
-        self.driver.find_element(By.LINK_TEXT, "Ir a Dataset 1").click()
+        self.driver.find_element(By.CSS_SELECTOR, ".sidebar-item:nth-child(9) .align-middle:nth-child(2)").click()
         self.driver.find_element(By.LINK_TEXT, "González, Juan Antonio").click()
         self.driver.find_element(By.CSS_SELECTOR, ".dropdown-item:nth-child(2)").click()
-
